message

+++ 0x02-functions_nested_loops/11-print_to_98.c
@@ -9,7 +9,7 @@ void print_to_98(int n)
 {
 	while (n <= 98)
 	{
-		_putchar(n);
+		_putchar((int)n);
 		n++;
 	}
 }
